feat: Reduce down to 1 web server
By using the Next.js static export mode, we can statically serve the build output from NestJS. This allows us to remove the Next.js server entirely, reducing resource usage and deployment complexity.

API_PORT has been removed in favour of UI_PORT, with NestJS now listening on this port as it's serving both the UI and API. The API port was mostly 'internal', with all API requests being proxied via Next.js (UI_PORT) to this port. The API_PORT environment variable was only introduced for those users who have another app listening on the same port. They are unaffected by this change. This is only a breaking change for users if they were calling the API via the API port, instead of using the /api/ route of the UI server. I've taken a bet that this is extremely uncommon.

The only thing we lose with this change is the Next.js image cache, but it's not entirely reliable and I think it doesn't really add much value.